@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "prisma generate && next build",
     "start": "next start",
     "lint": "next lint"
@@ -15,12 +15,12 @@
     "bootstrap-icons": "^1.11.3",
     "clsx": "^2.1.0",
     "moment": "^2.30.1",
-    "next": "14.1.0",
+    "next": "15.3.3",
     "postgresql": "^0.0.1",
     "primereact": "^10.5.1",
-    "react": "^18",
+    "react": "19.1.0",
     "react-bootstrap": "^2.10.1",
-    "react-dom": "^18",
+    "react-dom": "19.1.0",
     "react-icons": "^5.0.1",
     "reactstrap": "^9.2.2",
     "sharp": "^0.33.2",
@@ -30,7 +30,7 @@
   "devDependencies": {
     "autoprefixer": "^10.4.18",
     "eslint": "^8",
-    "eslint-config-next": "14.1.0",
+    "eslint-config-next": "15.3.3",
     "postcss": "^8.4.35",
     "prisma": "^5.11.0",
     "tailwindcss": "^3.4.1"
